+++ ai2html.js
@@ -348,7 +348,10 @@ if (runningInNode()) {
     readGitConfigFile,
     readYamlConfigFile,
     applyTemplate,
-    cleanText,
+    cleanHtmlText,
+    addEnclosingTag,
+    stripTag,
+    cleanCodeBlock,
     findHtmlTag,
     cleanHtmlTags,
     convertSettingsToYaml,
@@ -452,7 +455,7 @@ function render() {
   }
 
   // ================================================
-  // grab custom settings, html, css, js and text blocks
+  // import custom settings, html, css, js and text blocks
   // ================================================
   var documentHasSettingsBlock = false;
   var customBlocks = {};
@@ -474,11 +477,12 @@ function render() {
       parseSettingsEntries(entries, docSettings);
     } else if (type == 'text') {
       parseSettingsEntries(entries, docSettings);
-    } else { // custom js, css and html
+    } else {
+      // import custom js, css and html blocks
       if (!customBlocks[type]) {
         customBlocks[type] = [];
       }
-      customBlocks[type].push("\t\t" + cleanText(entries.join("\r\t\t")) + "\r");
+      customBlocks[type].push(cleanCodeBlock(type, entries.join('\r')));
     }
     if (objectOverlapsAnArtboard(thisFrame)) {
       hideTextFrame(thisFrame);
@@ -827,7 +831,7 @@ function makeKeyword(text) {
   return text.replace( /[^A-Za-z0-9_\-]+/g , "_" );
 }
 
-function cleanText(text) {
+function cleanHtmlText(text) {
   for (var i=0; i < htmlCharacterCodes.length; i++) {
     var charCode = htmlCharacterCodes[i];
     if (text.indexOf(charCode[0]) > -1) {
@@ -842,11 +846,15 @@ function straightenCurlyQuotesInsideAngleBrackets(text) {
   // typed into text blocks (Illustrator tends to automatically change single
   // and double quotes to curly quotes).
   // thanks to jashkenas
+  // var quoteFinder = /[\u201C‘’\u201D]([^\n]*?)[\u201C‘’\u201D]/g;
   var tagFinder = /<[^\n]+?>/g;
-  var quoteFinder = /[\u201C‘’\u201D]([^\n]*?)[\u201C‘’\u201D]/g;
   return text.replace(tagFinder, function(tag){
-    return tag.replace( /[\u201C\u201D]/g , '"' ).replace( /[‘’]/g , "'" );
+    return straightenCurlyQuotes(tag);
   });
+}
+
+function straightenCurlyQuotes(str) {
+  return str.replace( /[\u201C\u201D]/g , '"' ).replace( /[‘’]/g , "'" );
 }
 
 // Not very robust -- good enough for printing a warning
@@ -856,6 +864,24 @@ function findHtmlTag(str) {
     match = /<(\w+)[^>]*>/.exec(str);
   }
   return match ? match[1] : null;
+}
+
+function addEnclosingTag(tagName, str) {
+  var openTag = '<' + tagName;
+  var closeTag = '</' + tagName + '>';
+  if ((new RegExp(openTag)).test(str) === false) {
+    str = openTag + '>\r' + str;
+  }
+  if ((new RegExp(closeTag)).test(str) === false) {
+    str = str + '\r' + closeTag;
+  }
+  return str;
+}
+
+function stripTag(tagName, str) {
+  var open = new RegExp('<' + tagName + '[^>]*>', 'g');
+  var close = new RegExp('</' + tagName + '>', 'g');
+  return str.replace(open, '').replace(close, '');
 }
 
 // precision: number of decimals in rounded number
@@ -1256,6 +1282,23 @@ function initUtilityFunctions() {
   // Minified json2.js from https://github.com/douglascrockford/JSON-js
   // This code is in the public domain.
   if(typeof JSON!=="object"){JSON={}}(function(){"use strict";var rx_one=/^[\],:{}\s]*$/;var rx_two=/\\(?:["\\\/bfnrt]|u[0-9a-fA-F]{4})/g;var rx_three=/"[^"\\\n\r]*"|true|false|null|-?\d+(?:\.\d*)?(?:[eE][+\-]?\d+)?/g;var rx_four=/(?:^|:|,)(?:\s*\[)+/g;var rx_escapable=/[\\"\u0000-\u001f\u007f-\u009f\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/g;var rx_dangerous=/[\u0000\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/g;function f(n){return n<10?"0"+n:n}function this_value(){return this.valueOf()}if(typeof Date.prototype.toJSON!=="function"){Date.prototype.toJSON=function(){return isFinite(this.valueOf())?this.getUTCFullYear()+"-"+f(this.getUTCMonth()+1)+"-"+f(this.getUTCDate())+"T"+f(this.getUTCHours())+":"+f(this.getUTCMinutes())+":"+f(this.getUTCSeconds())+"Z":null};Boolean.prototype.toJSON=this_value;Number.prototype.toJSON=this_value;String.prototype.toJSON=this_value}var gap;var indent;var meta;var rep;function quote(string){rx_escapable.lastIndex=0;return rx_escapable.test(string)?'"'+string.replace(rx_escapable,function(a){var c=meta[a];return typeof c==="string"?c:"\\u"+("0000"+a.charCodeAt(0).toString(16)).slice(-4)})+'"':'"'+string+'"'}function str(key,holder){var i;var k;var v;var length;var mind=gap;var partial;var value=holder[key];if(value&&typeof value==="object"&&typeof value.toJSON==="function"){value=value.toJSON(key)}if(typeof rep==="function"){value=rep.call(holder,key,value)}switch(typeof value){case"string":return quote(value);case"number":return isFinite(value)?String(value):"null";case"boolean":case"null":return String(value);case"object":if(!value){return"null"}gap+=indent;partial=[];if(Object.prototype.toString.apply(value)==="[object Array]"){length=value.length;for(i=0;i<length;i+=1){partial[i]=str(i,value)||"null"}v=partial.length===0?"[]":gap?"[\n"+gap+partial.join(",\n"+gap)+"\n"+mind+"]":"["+partial.join(",")+"]";gap=mind;return v}if(rep&&typeof rep==="object"){length=rep.length;for(i=0;i<length;i+=1){if(typeof rep[i]==="string"){k=rep[i];v=str(k,value);if(v){partial.push(quote(k)+(gap?": ":":")+v)}}}}else{for(k in value){if(Object.prototype.hasOwnProperty.call(value,k)){v=str(k,value);if(v){partial.push(quote(k)+(gap?": ":":")+v)}}}}v=partial.length===0?"{}":gap?"{\n"+gap+partial.join(",\n"+gap)+"\n"+mind+"}":"{"+partial.join(",")+"}";gap=mind;return v}}if(typeof JSON.stringify!=="function"){meta={"\b":"\\b","\t":"\\t","\n":"\\n","\f":"\\f","\r":"\\r",'"':'\\"',"\\":"\\\\"};JSON.stringify=function(value,replacer,space){var i;gap="";indent="";if(typeof space==="number"){for(i=0;i<space;i+=1){indent+=" "}}else if(typeof space==="string"){indent=space}rep=replacer;if(replacer&&typeof replacer!=="function"&&(typeof replacer!=="object"||typeof replacer.length!=="number")){throw new Error("JSON.stringify")}return str("",{"":value})}}if(typeof JSON.parse!=="function"){JSON.parse=function(text,reviver){var j;function walk(holder,key){var k;var v;var value=holder[key];if(value&&typeof value==="object"){for(k in value){if(Object.prototype.hasOwnProperty.call(value,k)){v=walk(value,k);if(v!==undefined){value[k]=v}else{delete value[k]}}}}return reviver.call(holder,key,value)}text=String(text);rx_dangerous.lastIndex=0;if(rx_dangerous.test(text)){text=text.replace(rx_dangerous,function(a){return"\\u"+("0000"+a.charCodeAt(0).toString(16)).slice(-4)})}if(rx_one.test(text.replace(rx_two,"@").replace(rx_three,"]").replace(rx_four,""))){j=eval("("+text+")");return typeof reviver==="function"?walk({"":j},""):j}throw new SyntaxError("JSON.parse")}}})(); // jshint ignore:line
+}
+
+// Clean the contents of custom JS, CSS and HTML blocks and store them
+// (e.g. undo Illustrator's automatic quote conversion, where applicable)
+function cleanCodeBlock(type, raw) {
+  var clean = '';
+  if (type == 'html') {
+    clean = cleanHtmlText(straightenCurlyQuotesInsideAngleBrackets(raw));
+  } else if (type == 'js' ) {
+    // TODO: consider preserving curly quotes inside quoted strings
+    clean = straightenCurlyQuotes(raw);
+    clean = addEnclosingTag('script', clean);
+  } else if (type == 'css') {
+    clean = straightenCurlyQuotes(raw);
+    clean = stripTag('style', clean);
+  }
+  return clean;
 }
 
 function createSettingsBlock() {
@@ -1960,8 +2003,7 @@ function cleanHtmlTags(str) {
 }
 
 function generateParagraphHtml(pData, baseStyle, pStyles, cStyles) {
-  var classStr = '',
-      html, diff, range, text;
+  var html, diff, range, rangeHtml;
   if (pData.text.length === 0) { // empty pg
     // TODO: Calculate the height of empty paragraphs and generate
     // CSS to preserve this height (not supported by Illustrator API)
@@ -1970,20 +2012,19 @@ function generateParagraphHtml(pData, baseStyle, pStyles, cStyles) {
   diff = objectSubtract(pData.cssStyle, baseStyle);
   // Give the pg a class, if it has a different style than the base pg class
   if (diff) {
-    classStr = ' class="' + getTextStyleClass(diff, pStyles, 'pstyle') + '"';
+    html = '<p class="' + getTextStyleClass(diff, pStyles, 'pstyle') + '">';
+  } else {
+    html = '<p>';
   }
-  html = '<p' + classStr + '>';
   for (var j=0; j<pData.ranges.length; j++) {
     range = pData.ranges[j];
-    range.text = cleanHtmlTags(range.text);
+    rangeHtml = cleanHtmlText(cleanHtmlTags(range.text));
     diff = objectSubtract(range.cssStyle, pData.cssStyle);
     if (diff) {
-      html += '<span class="' + getTextStyleClass(diff, cStyles, 'cstyle') + '">';
+      rangeHtml = '<span class="' +
+      getTextStyleClass(diff, cStyles, 'cstyle') + '">' + rangeHtml + '</span>';
     }
-    html += cleanText(range.text);
-    if (diff) {
-      html += '</span>';
-    }
+    html += rangeHtml;
   }
   html += '</p>';
   return html;
@@ -3191,7 +3232,7 @@ function outputLocalPreviewPage(textForFile, localPreviewDestination, settings) 
 
 function addCustomContent(content, customBlocks) {
   if (customBlocks.css) {
-    content.css += "\r\t\t/* Custom CSS */\r\t\t" + customBlocks.css.join('\r\t\t') + '\r';
+    content.css += "\r\t/* Custom CSS */\r\t" + customBlocks.css.join('\r\t') + '\r';
     /*
     content = "\r\t<style type='text/css' media='screen,print'>\r" +
       "\t\t" + customBlocks.css.join('\r\t\t') +
@@ -3199,11 +3240,11 @@ function addCustomContent(content, customBlocks) {
     */
   }
   if (customBlocks.html) {
-    content.html += "\r\t<!-- Custom HTML -->\r" + customBlocks.html.join('\r') + '\r';
+    content.html += "\r<!-- Custom HTML -->\r" + customBlocks.html.join('\r') + '\r';
   }
   // TODO: assumed JS contained in <script> tag -- verify this?
   if (customBlocks.js) {
-    content.js += "\r\t<!-- Custom JS -->\r" + customBlocks.js.join('\r') + '\r';
+    content.js += "\r<!-- Custom JS -->\r" + customBlocks.js.join('\r') + '\r';
   }
 }