+++ test/ai2html-test.js
@@ -31,6 +31,45 @@ describe('Ai2html-specific functions', function () {
 
   })
 
+  describe('stripTag()', function() {
+    it('tests', function() {
+      assert.equal(lib.stripTag('style', '\r<style></style>'), '\r');
+      assert.equal(lib.stripTag('style', 'body {display: none}'), 'body {display: none}');
+
+    })
+  })
+
+  describe('addEnclosingTag()', function () {
+    it('tests', function() {
+      assert.equal(lib.addEnclosingTag('script', '\r<script></script>'), '\r<script></script>')
+      assert.equal(lib.addEnclosingTag('script', '<script type="text/javascript"></script>'), '<script type="text/javascript"></script>')
+      assert.equal(lib.addEnclosingTag('style', '\t<style>\r{body: margin: 0}\r</style> '), '\t<style>\r{body: margin: 0}\r</style> ')
+      assert.equal(lib.addEnclosingTag('script', '<script>'), '<script>\r</script>');
+      assert.equal(lib.addEnclosingTag('script', '</script>'), '<script>\r</script>');
+      assert.equal(lib.addEnclosingTag('script', ''), '<script>\r\r</script>');
+      assert.equal(lib.addEnclosingTag('style', ''), '<style>\r\r</style>');
+
+    })
+  })
+
+  describe('cleanCodeBlock()', function () {
+    it('straighten curly double quotes in JS', function () {
+      var str = lib.cleanCodeBlock('js', 'console.log(“hi”)');
+      assert.equal(str, '<script>\rconsole.log("hi")\r</script>');
+    })
+
+    it('straighten curly single quotes in JS', function () {
+      var str = lib.cleanCodeBlock('js', 'console.log(‘hi’)');
+      assert.equal(str, '<script>\rconsole.log(\'hi\')\r</script>');
+    })
+
+    it('straighten curly double quotes in HTML tags and convert to HTML in text', function () {
+      var str = lib.cleanCodeBlock('html', '<span class=”note”>“who’s idea?”</span>');
+      assert.equal(str, '<span class="note">&ldquo;who&rsquo;s idea?&rdquo;</span>');
+    })
+
+  })
+
   describe('cleanObjectName()', function () {
     it('remove colon-delimited annotation and convert spaces', function () {
       var name = lib.cleanObjectName("Layer 7:600,svg,label=Text Layer,height=400");